refactor: タスクの作成・更新に共通するオプション部分をmixinするように変更
